Move common/files to general/files
