+++ etsin_finder/frontend/js/components/qvain/files/ida/addItemsTreeItem.jsx
@@ -14,7 +14,7 @@ import {
   Icon,
   NoIcon,
   SmallLoader,
-} from '../../../common/files/items'
+} from '../../../general/files/items'
 
 const AddItemsTreeItemBase = ({ treeProps, item, level, parentArgs }) => {
   const { directoryView } = treeProps
