+++ etsin_finder/frontend/js/components/qvain/files/ida/selectedItemsTree.jsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types'
 import { inject, Observer } from 'mobx-react'
 
 import SelectedItemsTreeItem from './selectedItemsTreeItem'
-import { useRenderTree } from '../../../common/files/tree'
+import { useRenderTree } from '../../../general/files/tree'
 
 export function SelectedItemsTree(props) {
   const { Files } = props.Stores.Qvain
